# Avoid deep-copy when constructing groupby codes