- [x] Closes #9426
- [x] Benchmark added
- [x] User visible changes (including notable bug fixes) are documented in `whats-new.rst`

```
| Change   | Before [a8c98967]    | After [948366dc]    |   Ratio | Benchmark (Parameter)                            |
|----------|----------------------|---------------------|---------|--------------------------------------------------|
| -        | 117±0.6ms            | 14.8±0.2ms          |    0.13 | groupby.GroupByLongTime.time_mean(True, False)   |
| -        | 115±0.6ms            | 14.0±0.1ms          |    0.12 | groupby.GroupByLongTime.time_mean(False, False)  |
| -        | 113±0.9ms            | 9.23±0.3ms          |    0.08 | groupby.GroupByLongTime.time_mean(True, True)    |
| -        | 112±0.6ms            | 8.28±0.2ms          |    0.07 | groupby.GroupByLongTime.time_mean(False, True)   |
| -        | 105±0.5ms            | 3.45±0.03ms         |    0.03 | groupby.GroupByLongTime.time_setup(True, False)  |
| -        | 105±0.8ms            | 3.49±0.02ms         |    0.03 | groupby.GroupByLongTime.time_setup(True, True)   |
| -        | 105±0.9ms            | 2.40±0.03ms         |    0.02 | groupby.GroupByLongTime.time_setup(False, False) |
| -        | 105±1ms              | 2.45±0.01ms         |    0.02 | groupby.GroupByLongTime.time_setup(False, True)  |
```